+++ berkeley-function-call-leaderboard/bfcl/eval_checker/eval_runner_helper.py
@@ -1,175 +1,19 @@
-import json
 import os
-import re
 import statistics
 from pathlib import Path
-from typing import Union
 
 import numpy as np
 from bfcl._apply_function_credential_config import apply_function_credential_config
-from bfcl.constant import VERSION_PREFIX
 from bfcl.eval_checker.constant import *
 from bfcl.eval_checker.executable_eval.custom_exception import BadAPIStatusError
 from bfcl.eval_checker.model_metadata import *
-from bfcl.model_handler.handler_map import HANDLER_MAP
+from bfcl.utils import (
+    extract_test_category,
+    find_file_with_suffix,
+    load_file,
+    write_list_of_dicts_to_file,
+)
 from tqdm import tqdm
-
-
-def extract_test_category(input_string: Union[str, Path]) -> str:
-    input_string = str(input_string)
-    pattern = fr".*{VERSION_PREFIX}_(\w+?)(?:_score|_result)?\.json"
-    match = re.search(pattern, input_string)
-
-    # Check if there's a match and extract the captured group
-    if match:
-        return match.group(1)  # the first captured group (\w+)
-    else:
-        raise ValueError(f"Could not extract the test category from the input string: {input_string}")
-
-
-def find_file_with_suffix(folder_path: Path, suffix: str) -> Path:
-    for json_file in folder_path.glob("*.json"):
-        if extract_test_category(json_file) == suffix:
-            return json_file
-    raise FileNotFoundError(f"No JSON file found with suffix: {suffix}")
-
-def is_multi_turn(test_category):
-    return "multi_turn" in test_category
-
-def contain_multi_turn_irrelevance(test_category):
-    return "miss_func" in test_category or "miss_param" in test_category
-
-def is_executable(test_category):
-    return "exec" in test_category or "rest" in test_category
-
-
-def is_rest(test_category):
-    return "rest" in test_category
-
-
-def is_relevance_or_irrelevance(test_category):
-    return "relevance" in test_category or "irrelevance" in test_category
-
-
-def is_chatable(test_category):
-    return "chatable" in test_category
-
-
-def is_java(test_category):
-    return "java" in test_category
-
-
-def is_js(test_category):
-    return "javascript" in test_category
-
-
-def is_sql(test_category):
-    return "sql" in test_category
-
-
-def load_file(file_path):
-    result = []
-    with open(file_path) as f:
-        file = f.readlines()
-        for line in file:
-            result.append(json.loads(line))
-    return result
-
-
-def get_handler(model_name):
-    return HANDLER_MAP[model_name](model_name, temperature=0)  #Temperature doesn't matter for evaluation
-
-
-def write_list_of_dicts_to_file(filename, data, subdir=None):
-    if subdir:
-        # Ensure the subdirectory exists
-        os.makedirs(subdir, exist_ok=True)
-
-        # Construct the full path to the file
-        filename = os.path.join(subdir, filename)
-
-    # Write the list of dictionaries to the file in JSON format
-    with open(filename, "w") as f:
-        for i, entry in enumerate(data):
-            # Go through each key-value pair in the dictionary to make sure the values are JSON serializable
-            entry = make_json_serializable(entry)
-            json_str = json.dumps(entry)
-            f.write(json_str)
-            if i < len(data) - 1:
-                f.write("\n")
-
-
-def make_json_serializable(value):
-    if isinstance(value, dict):
-        # If the value is a dictionary, we need to go through each key-value pair recursively
-        return {k: make_json_serializable(v) for k, v in value.items()}
-    elif isinstance(value, list):
-        # If the value is a list, we need to process each element recursively
-        return [make_json_serializable(item) for item in value]
-    else:
-        # Try to serialize the value directly, and if it fails, convert it to a string
-        try:
-            json.dumps(value)
-            return value
-        except (TypeError, ValueError):
-            return str(value)
-
-
-def is_function_calling_format_output(decoded_output):
-    # Ensure the output is a list of dictionaries
-    if type(decoded_output) == list:
-        for item in decoded_output:
-            if type(item) != dict:
-                return False
-        return True
-    return False
-
-
-def is_executable_format_output(decoded_output):
-    # Ensure the output is a list of strings (one or more strings)
-    if type(decoded_output) == list:
-        if len(decoded_output) == 0:
-            return False
-        for item in decoded_output:
-            if type(item) != str:
-                return False
-        return True
-    return False
-
-
-def is_rest_format_output(decoded_output):
-    # Ensure the output is a list of one string
-    if type(decoded_output) == list:
-        if len(decoded_output) == 1 and type(decoded_output[0]) == str:
-            return True
-    return False
-
-
-def is_empty_output(decoded_output):
-    # This function is a patch to the ast decoder for relevance detection
-    # Sometimes the ast decoder will parse successfully, but the input doens't really have a function call
-    # [], [{}], and anything that is not in function calling format is considered empty (and thus should be marked as correct)
-    if not is_function_calling_format_output(decoded_output):
-        return True
-    if len(decoded_output) == 0:
-        return True
-    if len(decoded_output) == 1 and len(decoded_output[0]) == 0:
-        return True
-    return False
-
-
-def contain_unachievable_task(decoded_output: list[str]):
-    """
-    Check if the decoded output contains a function call to the `flag_task_unachievable` function, which signals that the task is unachievable in the given context.
-    The model should only invoke this function in the irrelevance scenario, and not in other scenarios.
-    
-    Args:
-        decoded_output (list[str]): The decoded output from the model, from the `decode_exec` function.
-    """
-    for item in decoded_output:
-        if "flag_task_unachievable" in item:
-            return True
-    return False
 
 
 def api_status_sanity_check_rest():
@@ -847,28 +691,3 @@ def update_leaderboard_table_with_score_file(leaderboard_table, score_path: Path
                     "accuracy": accuracy,
                     "total_count": total_count,
                 }
-
-
-def collapse_json_objects(file_path):
-    with open(file_path, "r") as file:
-        content = file.read()
-
-    objects = []
-    depth = 0
-    obj_start = 0
-    for i, char in enumerate(content):
-        if char == "{":
-            if depth == 0:
-                obj_start = i
-            depth += 1
-        elif char == "}":
-            depth -= 1
-            if depth == 0:
-                obj = content[obj_start : i + 1]
-                objects.append(obj)
-
-    with open(file_path, "w") as out_file:
-        for obj in objects:
-            json_obj = json.loads(obj)
-            compact_json = json.dumps(json_obj, separators=(",", ":"))
-            out_file.write(compact_json + "\n")
